move lti tests to seperate file
